fix: blog post title input stays after creation
when an action returns a redirect, it is thrown in the clientAction function. need to add a special case to rethrow Response objects in the catch block to not accidentally block the redirect request. a new header is also introduced in order to differentiate between a redirect to the new blog post vs other redirects so that the client can do some cleanup before the redirect (e.g. hiding the post title input)